Added Save All in Toolbar / Moved Save Code to Graph
